Refactor: standardize `Http` module paths to lowercase for consistency ♻️
- Renamed `Http` and `ApiResponse` files to `http` and `api-response` respectively.
- Updated related imports, stubs, and service provider mappings.
- Adjusted TypeScript contracts (e.g., `PaginationData` to `PaginationMetaData`) to reflect renamed paths.
