+++ resources/js/inertia/components/datatable/export-modal.tsx
@@ -2,7 +2,7 @@ import { TableSchema } from "@/components/datatable/types";
 import Button from "@/components/ui/button";
 import Modal from "@/components/ui/modal";
 import DownloadFile from "@/Hooks/DownloadFile";
-import Http from "@/Modules/Http/Http";
+import Http from "@/modules/http/http";
 import { FormEvent, useState } from "react";
 
 const ExportModal = ({
