+++ src/CubetaStarterServiceProvider.php
@@ -215,7 +215,7 @@ class CubetaStarterServiceProvider extends PackageServiceProvider
             __DIR__ . '/../resources/js/inertia/global.d.ts' => resource_path('/js/global.d.ts'),
             __DIR__ . '/../resources/js/inertia/vite-env.d.ts' => resource_path('/js/vite-env.d.ts'),
             __DIR__ . '/../resources/js/inertia/helper.ts' => resource_path('/js/helper.ts'),
-            __DIR__ . '/../resources/js/inertia/Modules' => resource_path('/js/Modules'),
+            __DIR__ . '/../resources/js/inertia/modules' => resource_path('/js/modules'),
             __DIR__ . '/../resources/css/inertia' => resource_path('/css'),
             __DIR__ . '/../public/images' => public_path('/images'),
